client/requirements.txt
grpcio==1.36.1
grpcio-tools==1.36.1
kafka-python==2.0.2
elasticsearch==7.13.4
protobuf==3.19.1
